+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## 0.2.2
+
+### Hints
+
+- Removed ReadXplorer from module development since RX-CLI is deprecated
+
+### Added (3 changes)
+
+- Added mamba support and made mamba default when using conda
+- Added support of gff.gz files (DGE Analysis Module)
+- Added 'N' (None) option at module selection (Curare Wizard)
+
+### Fixed (4 changes)
+
+- Fixed bug when all options in a module are optional
+- Fixed parsing errors of pipeline and samples file
+- Check and create missing output folders
+- Use curare snakefiles folder as default (Curare Wizard)
+
+### Changed (1 change)
+
+- Changed '--attribute' list from space-seperated to comma-seperated (DGE Analysis Module)
+
 ## 0.2.1